requirements.txt
openai>=1.0.0
pypdf>=3.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy
deskew
dateparser
openpyxl>=3.1.0